refactor: extract notification tracking payload mappers
