+++ packages/pn-personagiuridica-webapp/src/pages/Notifiche.page.tsx
@@ -6,14 +6,11 @@ import {
   A11yMessage,
   ApiErrorWrapper,
   CustomPagination,
-  Notification,
   NotificationColumnData,
-  NotificationStatus,
   PaginationData,
   Sort,
   TitleBox,
   calculatePages,
-  isNewNotification,
   useEventEmitter,
   useHasPermissions,
   useIsMobile,
@@ -24,7 +21,6 @@ import LoadingPageWrapper from '../components/LoadingPageWrapper/LoadingPageWrap
 import DesktopNotifications from '../components/Notifications/DesktopNotifications';
 import GroupSelector from '../components/Notifications/GroupSelector';
 import MobileNotifications from '../components/Notifications/MobileNotifications';
-import { PGEventPayloads } from '../models/PGEventPayloads';
 import { PGEventsType } from '../models/PGEventsType';
 import { PNRole } from '../models/User';
 import { ContactSource } from '../models/contacts';
@@ -33,41 +29,11 @@ import { setNotificationFilters, setPagination, setSorting } from '../redux/dash
 import { useAppDispatch, useAppSelector } from '../redux/hooks';
 import { RootState } from '../redux/store';
 import PGEventStrategyFactory from '../utility/MixpanelUtils/PGEventStrategyFactory';
+import { mapNotificationListToEventPayload } from '../utility/MixpanelUtils/mappers/notificationPayloadMappers';
 
 type Props = {
   isDelegatedPage?: boolean;
 };
-
-const buildYourNotificationTrackingPayload = (
-  notifications: Array<Notification>,
-  pageNumber: number,
-  domicileBannerType?: string
-): PGEventPayloads[PGEventsType.SEND_PG_YOUR_NOTIFICATION] => ({
-  page_number: pageNumber,
-  total_count: notifications.length,
-  unread_count: notifications.filter((notification) =>
-    isNewNotification(notification.notificationStatus)
-  ).length,
-  delivered_count: notifications.filter(
-    (notification) => notification.notificationStatus === NotificationStatus.DELIVERED
-  ).length,
-  opened_count: notifications.filter(
-    (notification) => notification.notificationStatus === NotificationStatus.VIEWED
-  ).length,
-  expired_count: notifications.filter(
-    (notification) => notification.notificationStatus === NotificationStatus.EFFECTIVE_DATE
-  ).length,
-  not_found_count: notifications.filter(
-    (notification) => notification.notificationStatus === NotificationStatus.UNREACHABLE
-  ).length,
-  cancelled_count: notifications.filter(
-    (notification) => notification.notificationStatus === NotificationStatus.CANCELLED
-  ).length,
-  effective_date_count: notifications.filter(
-    (notification) => notification.notificationStatus === NotificationStatus.EFFECTIVE_DATE
-  ).length,
-  ...(domicileBannerType && { banner: domicileBannerType }),
-});
 
 const Notifiche = ({ isDelegatedPage = false }: Props) => {
   const dispatch = useAppDispatch();
@@ -132,7 +98,7 @@ const Notifiche = ({ isDelegatedPage = false }: Props) => {
         if (!isDelegatedPage) {
           PGEventStrategyFactory.triggerEvent(
             PGEventsType.SEND_PG_YOUR_NOTIFICATION,
-            buildYourNotificationTrackingPayload(
+            mapNotificationListToEventPayload(
               data.resultsPage,
               pagination.page,
               domicileBannerTypeRef.current
